telegram_bot/requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
openai>=1.0.0
httpx==0.25.0
aiosqlite==0.19.0
pydantic==2.5.2
google-cloud-speech==2.24.1
mutagen==1.47.0
redis==5.0.1
